Fix new clippy lints

+++ gio/src/pollable_input_stream.rs
@@ -120,7 +120,7 @@ impl<O: IsA<PollableInputStream>> PollableInputStreamExtManual for O {
     ) -> Result<isize, glib::Error> {
         let cancellable = cancellable.map(|c| c.as_ref());
         let gcancellable = cancellable.to_glib_none();
-        let count = buffer.len() as usize;
+        let count = buffer.len();
         unsafe {
             let mut error = ptr::null_mut();
             let ret = ffi::g_pollable_input_stream_read_nonblocking(
